Change positions and rename files
