+++ src/app/layout.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from "next";
 import { Dongle } from "next/font/google";
-import "./globals.css";
+import "../styles/globals.css";
 import Link from "next/link";
 import { signOut } from "./features/auth/actions/sign-out";
 import { getAuth } from "./features/auth/queries/get-auth";
