+++ src/components/home/index.ts
@@ -1,5 +1,5 @@
 import Card from "./card";
 import Review from "./reviews";
-import CustomSwiper from "./swiper";
+import CustomSwiper from "../CustomSwiper";
 
-export {Review, Card, CustomSwiper}
+export { Review, Card, CustomSwiper };
