+++ src/components/index.ts
@@ -1,5 +1,6 @@
+import CustomSwiper from "./CustomSwiper";
 import HorizontalRuleStar from "./HorizontalRuleStar";
 import NewsletterInput from "./NewsletterInput";
 import Video from "./Video";
 
-export { HorizontalRuleStar, NewsletterInput, Video };
+export { CustomSwiper, HorizontalRuleStar, NewsletterInput, Video };
